ERC721 auction part updated
