+++ src/components/modals/placebid-modal/index.jsx
@@ -16,168 +16,8 @@ import TradeContract from "../../../contracts/json/trade.json";
 import TransferProxy from "../../../contracts/json/TransferProxy.json";
 import TokenContract from "../../../contracts/json/ERC20token.json";
 
-const PlaceBidModal = ({ show, handleModal, product }) => {
+const PlaceBidModal = ({ show, handleModal, product, handleSubmit }) => {
     const { walletData, setWalletData } = useContext(WalletData);
-
-    const router = useRouter();
-
-    async function switchNetwork(chainId) {
-        if (
-            parseInt(window.ethereum.networkVersion, 2) === parseInt(chainId, 2)
-        ) {
-            console.log(`Network is already with chain id ${chainId}`);
-            return true;
-        }
-        try {
-            const res = await window.ethereum.request({
-                method: "wallet_switchEthereumChain",
-                params: [{ chainId }],
-            });
-            // console.log(res);
-            return true;
-        } catch (switchError) {
-            // console.log(switchError);
-            toast.error("Failed to change the network.");
-        }
-        return false;
-    }
-
-    async function completeAuction() {
-        // update collectible putOnSale, saleType to true
-        const response = await axios.put(
-            `${process.env.NEXT_PUBLIC_STRAPI_API_URL}/api/collectibles/${product.id}`,
-            {
-                data: {
-                    putOnSale: false,
-                    owner: walletData.account,
-                },
-            }
-        );
-        console.log(response);
-    }
-
-    async function StoreData(price) {
-        try {
-            const signer = walletData.provider.getSigner();
-            if (product.collection.data.collectionType === "Single") {
-                const { contractAddress } = product.collection.data;
-                // Pull the deployed contract instance
-                const tokenContract = new walletData.ethers.Contract(
-                    TokenContract.address,
-                    TokenContract.abi,
-                    signer
-                );
-
-                const allowance = await tokenContract.allowance(
-                    walletData.account,
-                    TransferProxy.address
-                );
-                const allowanceAmount = parseInt(allowance._hex, 16);
-                console.log(allowanceAmount, parseInt(price));
-                if (allowanceAmount < parseInt(price)) {
-                    // approve nft first
-                    const transaction = await tokenContract.approve(
-                        TransferProxy.address,
-                        parseInt(price)
-                    );
-                    const receipt = await transaction.wait();
-                    console.log(receipt);
-                }
-                if (product.auction.data.sellType == "FixedPrice") {
-                    const seller = product.auction.data.walletAddress;
-                    const buyer = walletData.account;
-                    const erc20Address = TokenContract.address;
-                    const nftAddress = contractAddress;
-                    const nftType =
-                        product.collection.data.collectionType === "Single"
-                            ? 1
-                            : 0;
-                    const unitPrice = "1";
-                    const skipRoyalty = true;
-                    const amount = `${price}`;
-                    const tokenId = `${product.nftID}`;
-                    const tokenURI = "";
-                    const supply = `${1}`;
-                    const royaltyFee = `${10}`;
-                    const qty = `${1}`;
-
-                    // Pull the deployed contract instance
-                    const tradeContract = new walletData.ethers.Contract(
-                        TradeContract.address,
-                        TradeContract.abi,
-                        signer
-                    );
-
-                    console.log("Calling direct buy");
-                    const transaction = await tradeContract.buyAsset([
-                        seller,
-                        buyer,
-                        erc20Address,
-                        nftAddress,
-                        nftType,
-                        unitPrice,
-                        skipRoyalty,
-                        amount,
-                        tokenId,
-                        tokenURI,
-                        supply,
-                        royaltyFee,
-                        qty,
-                    ]);
-                    const receipt = await transaction.wait();
-                    console.log(receipt);
-                    if (receipt) {
-                        completeAuction();
-                    }
-                }
-            } else if (product.collection.data.collectionType === "Multiple") {
-                const contractAddress =
-                    product.collection.data.contractAddress1155;
-            }
-
-            // create bidding
-            const res = await axios.post(
-                `${process.env.NEXT_PUBLIC_STRAPI_API_URL}/api/biddings`,
-                {
-                    data: {
-                        bidPrice: price,
-                        bidderAddress: walletData.account,
-                        timeStamp: new Date(),
-                        auction: product.auction.data.id,
-                    },
-                }
-            );
-            // console.log(res);
-            if (product.auction.data.sellType === "FixedPrice") {
-                toast.success("NFT purchased successfully!");
-            } else {
-                toast.success("Bidding placed successfully!");
-            }
-            router.reload();
-        } catch (error) {
-            console.log(error);
-        }
-    }
-
-    const handleSubmit = (event) => {
-        event.preventDefault();
-        if (!walletData.isConnected) {
-            toast.error("Please connect wallet first");
-            return;
-        } // chnage network
-        if (product.collection.data.networkType === "Ethereum") {
-            if (!switchNetwork(ETHEREUM_NETWORK_CHAIN_ID)) {
-                // ethereum testnet
-                return;
-            }
-        } else if (product.collection.data.networkType === "Polygon") {
-            if (!switchNetwork(POLYGON_NETWORK_CHAIN_ID)) {
-                // polygon testnet
-                return;
-            }
-        }
-        StoreData(event.target.price.value);
-    };
 
     return (
         <Modal
